Upload function delete, edit article, comment and replycomment

+++ SuViet/src/main/java/com/example/SuViet/repository/RepliesCommentRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.SuViet.model.RepliesComment;
 
-public interface RepliesCommentRepository extends JpaRepository<RepliesComment, Long> {
+public interface RepliesCommentRepository extends JpaRepository<RepliesComment, Integer> {
 }
